Vectorize linalg.fractional_matrix_power, and make it consistent with dtype
