+++ geomstats/_backend/pytorch/linalg.py
@@ -6,6 +6,11 @@ import torch as _torch
 
 from .._backend_config import np_atol as atol
 from ..numpy import linalg as _gsnplinalg
+from ._common import cast as _cast
+from ._dtype import get_default_cdtype as _get_default_cdtype
+from ._dtype import get_default_dtype as _get_default_dtype
+from ._dtype import is_complex as _is_complex
+from ._dtype import is_floating as _is_floating
 
 
 class _Logm(_torch.autograd.Function):
@@ -135,9 +140,20 @@ def is_single_matrix_pd(mat):
         return False
 
 
-def fractional_matrix_power(mat, power):
+def fractional_matrix_power(A, t):
     """Compute the fractional power of a matrix."""
-    mat = mat.numpy()
-    mat_power = _scipy.linalg.fractional_matrix_power(mat, power)
-    mat_power = _torch.tensor(mat_power)
-    return mat_power
+    A = A.numpy()
+    if A.ndim == 2:
+        out = _scipy.linalg.fractional_matrix_power(A, t)
+    else:
+        out = _np.stack([_scipy.linalg.fractional_matrix_power(A_, t) for A_ in A])
+    out = _torch.tensor(out)
+
+    if _is_floating(out):
+        if out.dtype != _get_default_dtype():
+            return _cast(out, _get_default_dtype())
+    elif _is_complex(out):
+        if out.dtype != _get_default_cdtype():
+            return _cast(out, _get_default_cdtype())
+
+    return out
